requirements.txt
certifi==2019.11.28
numpy==1.18.1
Pillow==7.0.0
